int codepoints -> Uchar.t codepoints
These fixes are preliminary. There's a need for more tests, and the
change made to the code generated by ppx_sedlex to handle the EOF case
needs cleanup, as right now it generates duplicate code instead of
creating an inner function.

Still to do: possibly removing the locally maintained unicode handling
code and replacing it with parts of dbuenzli's unicode libraries or
some equivalent.